Added admin access while delete/add/update

+++ src/main/java/com/personalproject/studentinformationportal/service/StudentInfoService.java
@@ -8,6 +8,7 @@ import com.personalproject.studentinformationportal.model.DeleteResponse;
 import com.personalproject.studentinformationportal.model.UpdateResponse;
 import io.micrometer.common.util.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 
 import com.personalproject.studentinformationportal.model.StudentInfo;
@@ -17,7 +18,10 @@ import org.springframework.util.ObjectUtils;
 public class StudentInfoService {
 	@Autowired
 	private StudentInfoRepository studentInfoRepository;
-	
+
+	@Value("${x-admin-key}")
+	private String adminKey;
+
 	public StudentInfo addStudent(StudentInfo studentInfo) {
 		return studentInfoRepository.save(studentInfo);
 	}
@@ -30,44 +34,53 @@ public class StudentInfoService {
 		return studentInfoRepository.findById(id);
 	}
 
-	public DeleteResponse deleteStudent(Integer studentId) {
+	public DeleteResponse deleteStudent(Integer studentId, String adminKeyInput) {
+		DeleteResponse deleteResponse = new DeleteResponse();
 		Object response = studentInfoRepository.findById(studentId);
-		if (((Optional) response).isEmpty() ) {
-			return new DeleteResponse("", "Request not succeeded as student with "+ studentId+" does not exist") ;
+		if (adminKeyInput == null || !adminKeyInput.equalsIgnoreCase(adminKey)) {
+			deleteResponse.setFailedMessage("Invalid admin key or not authorized to add student");
+		} else {
+			if (((Optional) response).isEmpty() ) {
+				return new DeleteResponse("", "Request not succeeded as student with "+ studentId+" does not exist") ;
+			}
+			studentInfoRepository.deleteById(studentId);
+			deleteResponse.setSuccessMessage("Successfully deleted student with "+ studentId+" from database"); ;
 		}
-		studentInfoRepository.deleteById(studentId);
-		return new DeleteResponse("Successfully deleted student with "+ studentId+" from database", "Student deleted successfully") ;
-
+		return deleteResponse;
 	}
 
 	public List<StudentInfo> searchStudentByName(String name) {
 		return studentInfoRepository.findByName(name);
 	}
 
-	public UpdateResponse updateStudent(int id, StudentInfo updatedStudentInfo) {
+	public UpdateResponse updateStudent(int id, StudentInfo updatedStudentInfo, String adminKeyInput) {
 		UpdateResponse response = new UpdateResponse();
-		Optional<StudentInfo> existingStudentOptional = studentInfoRepository.findById(id);
-		if(existingStudentOptional.isPresent()) {
-			if(!StringUtils.isEmpty(updatedStudentInfo.getName())) {
-				existingStudentOptional.get().setName(updatedStudentInfo.getName());
-			}
-			if (!StringUtils.isEmpty(updatedStudentInfo.getFatherName())) {
-				existingStudentOptional.get().setFatherName(updatedStudentInfo.getFatherName());
-			}
-			if (!ObjectUtils.isEmpty(updatedStudentInfo.getAge())) {
-				existingStudentOptional.get().setAge(updatedStudentInfo.getAge());
-			}
-			if (!StringUtils.isEmpty(updatedStudentInfo.getMotherName())) {
-				existingStudentOptional.get().setMotherName(updatedStudentInfo.getMotherName());
-			}
-			if(!StringUtils.isEmpty(updatedStudentInfo.getAddress())) {
-				existingStudentOptional.get().setAddress(updatedStudentInfo.getAddress());
-			}
-		};
-		studentInfoRepository.save(existingStudentOptional.get());
-		response.setSuccessMessage("Student updated successfully");
-		response.setStudentInfo(studentInfoRepository.findById(id).get());
 
+		if (adminKeyInput == null || !adminKeyInput.equalsIgnoreCase(adminKey)) {
+			response.setFailedMessage("Invalid admin key or not authorized to add student");
+		} else {
+			Optional<StudentInfo> existingStudentOptional = studentInfoRepository.findById(id);
+			if(existingStudentOptional.isPresent()) {
+				if(!StringUtils.isEmpty(updatedStudentInfo.getName())) {
+					existingStudentOptional.get().setName(updatedStudentInfo.getName());
+				}
+				if (!StringUtils.isEmpty(updatedStudentInfo.getFatherName())) {
+					existingStudentOptional.get().setFatherName(updatedStudentInfo.getFatherName());
+				}
+				if (!ObjectUtils.isEmpty(updatedStudentInfo.getAge())) {
+					existingStudentOptional.get().setAge(updatedStudentInfo.getAge());
+				}
+				if (!StringUtils.isEmpty(updatedStudentInfo.getMotherName())) {
+					existingStudentOptional.get().setMotherName(updatedStudentInfo.getMotherName());
+				}
+				if(!StringUtils.isEmpty(updatedStudentInfo.getAddress())) {
+					existingStudentOptional.get().setAddress(updatedStudentInfo.getAddress());
+				}
+			};
+			studentInfoRepository.save(existingStudentOptional.get());
+			response.setSuccessMessage("Student updated successfully");
+			response.setStudentInfo(studentInfoRepository.findById(id).get());
+		}
 		return response;
 	}
 }
